requirements.txt
absl-py==2.1.0
accelerate==0.29.2
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==4.1.0
appdirs==1.4.4
argon2-cffi==21.1.0
arrow==1.3.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Automat==20.2.0
Babel==2.13.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.4
bottle==0.12.19
Bottleneck==1.3.2
Brotli==1.0.9
cachetools==5.0.0
certifi==2024.2.2
cffi==1.15.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloud-init==23.3.3
colorama==0.4.4
comm==0.2.0
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
ctop==1.0.0
cycler==0.11.0
dacite==1.8.1
datasets==2.18.0
dbus-python==1.2.18
debugpy==1.8.0
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
docker==5.0.3
entrypoints==0.4
et-xmlfile==1.0.1
evaluate==0.4.1
exceptiongroup==1.2.0
fastjsonschema==2.19.0
filelock==3.13.4
flake8==4.0.1
flatbuffers===1.12.1-git20200711.33e2d80-dfsg1-0.6
fonttools==4.29.1
fqdn==1.5.1
frozenlist==1.4.1
fs==2.4.12
fsspec==2024.2.0
future==0.18.2
gast==0.5.2
Glances==3.2.4.2
google-auth==1.5.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.30.2
h5py==3.6.0
h5py.-debian-h5py-serial==3.6.0
html5lib==1.1
htmlmin==0.1.12
httplib2==0.20.2
huggingface-hub==0.22.2
hyperlink==21.0.0
icdiff==2.0.4
idna==3.6
ImageHash==4.3.1
importlib-metadata==4.6.4
incremental==21.3.0
influxdb==5.3.1
iniconfig==1.1.1
iotop==0.6
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jax==0.4.14
jaxlib==0.4.14
jdcal==1.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.14
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-console==6.4.0
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter-ydoc==1.1.1
jupyter_client==8.6.0
jupyter_collaboration==1.2.0
jupyter_core==5.5.0
jupyter_server==2.12.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-pygments==0.1.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.2
kaptan==0.5.12
keras==2.13.1
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libtmux==0.10.1
llvmlite==0.41.1
lxml==4.8.0
lz4==3.1.3+dfsg
Markdown==3.3.6
MarkupSafe==2.1.5
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==3.0.2
ml-dtypes==0.2.0
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.0.5
multimethod==1.10
multiprocess==0.70.16
nbclient==0.5.6
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==3.3
nltk==3.8.1
nose==1.3.7
notebook==6.4.8
notebook_shim==0.2.3
numba==0.58.1
numexpr==2.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
odfpy==1.4.2
olefile==0.46
openpyxl==3.0.9
opt-einsum==3.3.0
overrides==7.4.0
packaging==24.0
pandas==2.2.1
pandas-profiling==3.6.6
pandocfilters==1.5.0
parso==0.8.1
patsy==0.5.4
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
pluggy==0.13.0
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==4.21.12
psutil==5.9.8
ptyprocess==0.7.0
py==1.10.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.11.0
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pysmi==0.3.2
pysnmp==4.4.12
pystache==0.6.0
pytest==6.2.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-json-logger==2.0.7
python-magic==0.4.24
pythran==0.10.0
pytz==2024.1
PyWavelets==1.5.0
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.31.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rouge-score==0.1.2
rpds-py==0.13.2
rsa==4.8
safetensors==0.4.2
scikit-learn==0.23.2
scipy==1.8.0
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.2
sentencepiece==0.2.0
service-identity==18.1.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
sos==4.5.6
soupsieve==2.3.1
ssh-import-id==5.11
statsmodels==0.14.0
sympy==1.12
systemd-python==234
tables==3.7.0
tangled-up-in-unicode==0.2.0
tensorboard==2.13.0
tensorflow==2.13.1
tensorflow-estimator==2.13.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tmuxp==1.9.2
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
torch==2.2.2
torchvision==0.15.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.0
transformers==4.39.3
triton==2.2.0
Twisted==22.1.0
typeguard==4.1.5
types-python-dateutil==2.8.19.14
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.3.0
urllib3==2.2.1
virtualenv==20.13.0+ds
visions==0.7.5
wadllib==1.3.6
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
widgetsnbextension==4.0.9
wordcloud==1.9.2
wrapt==1.13.3
xlwt==1.3.0
xxhash==3.4.1
y-py==0.6.2
yarl==1.9.4
ydata-profiling==4.6.3
ypy-websocket==0.12.4
zipp==1.0.0
zope.interface==5.4.0
